Can now Get ID from Company. Had to change ID types to long instead of int

+++ WindowsApp/Core/Interfaces/iSqlServerDataAccess.cs
@@ -8,7 +8,7 @@ namespace Core.Interfaces
 {
     public interface iSqlServerDataAccess
     {
-        int ExecuteQuery_SingleID(string sql);
+        long ExecuteQuery_SingleID(string sql);
         int ExecuteQuery_NoReturnType(string sql);
         TruckDriver ExecuteQuery_SingleTD(string sql);
         LumberAssociate ExecuteQuery_SingleLA(string sql);
